fix: pin dependencies to their minor versions in accordance with semver
This causes friction for downstream users that try to upgrade, since the
pinning here is overly strict; since packages on Hex are generally
expected to respect semantic versioning, minor version upgrades should
not get blocked by dependencies that pin to the patch version. It is
possible to override dependencies when necessary, but that is a heavy
hammer that should only be used in specific circumstances, and not the
default solution/recommendation of a project.